requirements.txt
flask
pandas
gspread
google-auth
openai
apscheduler